Calculate means and standard deviations

Under construction: this function is not working yet

Usage

E(
  Ynames,
  X = NULL,
  learnt,
  quantiles = c(5, 95)/100,
  nsamples = 100L,
  parallel = TRUE,
  silent = FALSE,
  usememory = TRUE
)

Arguments

X

matrix or data.table or NULL: set of values of variates on which we want to condition the joint probability of Y. If NULL (default), no conditioning is made (except for conditioning on the learning dataset and prior assumptions). One variate per column, one set of values per row.

learnt

Either a string with the name of a directory or full path for an 'learnt.rds' object, or such an object itself

quantiles

numeric vector, between 0 and 1, or NULL: desired quantiles of the variability of the probability for Y. Default c(0.05, 0.95) or the 5% and 95% quantiles.

nsamples

integer or NULL: desired number of samples of the variability of the probability for Y. Default 100.

parallel

logical or integer: whether to use pre-existing parallel workers, or how many to create and use. Default TRUE.

silent

logical: give warnings or updates in the computation? Default FALSE.

usememory

logical: save partial results to disc, to avoid crashes? Default TRUE.

Y1names

String vector: joint variates the joint probability of. One variate per column, one set of values per row.

Value

A list of: (1) a matrix with the probabilities P(Y|X,data,assumptions), for all combinations of values of Y (rows) and X (columns); (2) an array with the variability quantiles (3rd dimension of the array) for such probabilities; (3) an array with the variability samples (3rd dimension of the array) for such probabilities.