Calculate mutual information between groups of joint variates

Calculate mutual information between groups of joint variates

Usage

mutualinfo(
  Y1names,
  Y2names,
  X = NULL,
  learnt,
  nsamples = 3600,
  unit = "Sh",
  parallel = TRUE,
  silent = FALSE
)

Arguments

Y1names

String vector: first group of joint variates

Y2names

String vector or NULL: second group of joint variates

X

matrix or data.frame or NULL: values of some variates conditional on which we want the probabilities

learnt

Either a string with the name of a directory or full path for an 'learnt.rds' object, or such an object itself

nsamples

numeric: number of samples from which to approximately calculate the mutual information. Default 3600

unit

Either one of 'Sh' (default), 'Hart', 'nat', or a positive real

parallel,

logical or numeric: whether to use pre-existing parallel workers, or how many to create and use

silent

logical: give warnings or updates in the computation

Value

A list with the mutual information, its error, and its unit