Plot an object of class "probability"

This plot method is a utility to plot probabilities obtained with Pr or tailPr, as well as their uncertainties. The probabilities are plotted either against Y, with one curve for each value of X, or vice versa.

Usage

# S3 method for class 'probability'
plot(
  p,
  variability = NULL,
  PvsY = NULL,
  legend = TRUE,
  lwd = 3,
  lty = 1:5,
  col = palette(),
  xlab = NULL,
  ylab = "probability",
  ylim = c(0, NA),
  add = FALSE,
  ...
)

Arguments

p

Object of class "probability", obtained with Pr or tailPr.

variability

One of the values "quantiles", "samples", "none" (equivalent to NA or FALSE), or NULL (default), in which case the variability available in p is used. This argument chooses how to represent the variability of the probability; see Pr. If the requested variability is not available in the object p, then a warning is issued and no variability is plotted.

PvsY

Logical or NULL: should probabilities be plotted against their Y argument? If NULL, the argument between Y and X having larger number of values is chosen. As many probability curves will be plotted as the number of values of the other argument.

legend

Logical: plot a legend of the different curves?

...

Other parameters to be passed to matplot.