Calculate posterior probabilities

Calculate posterior probabilities

Usage

samplesFDistribution(
  Y,
  X,
  learnt,
  subsamples,
  jacobian = TRUE,
  fn = NULL,
  combine = rbind,
  parallel = TRUE,
  silent = FALSE
)

Arguments

Y

matrix or data.table: values of some variates of which we want the joint probability one variate per column

X

matrix or data.table: values of some variates conditional on which we want the joint probability one variate per column

learnt

Either a string with the name of a directory or full path for an 'learnt.rds' object, or such an object itself

subsamples

numeric: number of Monte Carlo samples to use

jacobian

include the Jacobian in the output probability

fn

NULL or function to apply to the group of MCsamples, example 'identity' or 'mean'

combine

how to combine the output for the different variate values

parallel

logical or numeric: whether to use pre-existing parallel workers, or how many to create and use

silent

logical: give warnings or updates in the computation

Value

The frequencies F(Y|X) corresponding to the Monte Carlo samples