Plot probabilities

Utility function to plot probabilities obtained with Pr. The advantage of this function over plot is that its x argument can be a vector of character values, as typically the case for nominal or ordinal variates (see metadata). The plot will have these characters as labels of the x-axis.

Usage

tplot(
  x,
  y,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  asp = NA,
  n = 10,
  family = "",
  xticks = NULL,
  xlabels = TRUE,
  yticks = NULL,
  ylabels = TRUE,
  cex = 1.5,
  ly = NULL,
  lx = NULL,
  mar = NULL,
  lty.axis = 1,
  lwd.axis = 0,
  lwd.ticks = 1,
  col.ticks = "#bbbbbb80",
  col.lab = "black",
  cex.axis = 1.12,
  las.y = 1,
  xgrid = NULL,
  ygrid = NULL,
  main = NULL,
  cex.main = 1.5,
  xlab = NULL,
  ylab = NULL,
  cex.lab = 1.5,
  type = "l",
  col = palette(),
  pch = c(1, 0, 2, 5, 6, 3, 4),
  lty = 1:4,
  lwd = 2,
  alpha = NA,
  border = palette(),
  border.alpha = NA,
  xtransf = NULL,
  ytransf = NULL,
  add = FALSE
)

Arguments

x

Numeric or character: Set of values that were used as Y argument to Pr (typically they make sense if Y only was a single variate rather than aconjunction of several).

y

Numeric: a vector of same length as x, or a matrix having as many rows as x. Typically this is a vector of probabilities obtained with Pr, as its $values value, on as its $samples value.